@@ -1,10 +1,13 @@
 package org.example;
 
-import org.junit.Test;
+
+import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+;
 
 public class Exercise10Test {
     @Test
