@@ -1,13 +1,14 @@
 package org.example;
 
-import org.junit.Test;
+
+import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 
 import static org.example.Exercise2.countImpl1;
 import static org.example.Exercise2.countImpl2;
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class Exercise2Test {
 
