@@ -2,10 +2,11 @@ package org.example;
 
 import org.example.exercise4.LoadBalancer;
 import org.example.exercise4.Node;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
 
 public class Exercise4Test {
     @Test
